Split tests to avoid huge class
